requirements.txt
openai==1.12.0
python-dotenv==1.0.1
googlesearch-python==1.2.3
beautifulsoup4==4.12.3
requests==2.31.0
pyautogui==0.9.54
spotipy==2.23.0
swarm==0.1.0